Move hyperoptresolver to resolvers package
