+++ freqtrade/resolvers/__init__.py
@@ -1,2 +1,3 @@
 from freqtrade.resolvers.iresolver import IResolver  # noqa: F401
-from freqtrade.resolvers.strategyresolver import StrategyResolver  # noqa: F401
+from freqtrade.resolvers.hyperopt_resolver import HyperOptResolver  # noqa: F401
+from freqtrade.resolvers.strategy_resolver import StrategyResolver  # noqa: F401
